% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.parobs.R
\name{bayes.parobs}
\alias{bayes.parobs}
\title{Fit Bayesian Inference for Multivariate Meta-Regression With a Partially Observed Within-Study Sample Covariance Matrix}
\usage{
bayes.parobs(Outcome, SD, XCovariate, WCovariate, Treat, Trial, Npt,
  fmodel = 1, prior = list(), mcmc = list(), control = list(),
  verbose = FALSE)
}
\arguments{
\item{Outcome}{aggregate mean of the responses for each arm of each study}

\item{SD}{standard deviation of the responses for each arm of each study}

\item{XCovariate}{aggregate covariates for the fixed effects}

\item{WCovariate}{aggregate covariates for the random effects}

\item{Treat}{treatment applied; equivalent to the arm number of each study; the number of unique treatments must be equal across trials}

\item{Trial}{trial identifier}

\item{Npt}{number of observations per trial}

\item{fmodel}{the model number; defaults to M1}

\item{prior}{list of hyperparameters; when not given, algorithm will run in default setting}

\item{mcmc}{list of MCMC-related parameters: number of burn-ins (ndiscard), number of thinning(nskip), and posterior sample size (nkeep)}

\item{control}{list of parameters for localized Metropolis algorithm: the step sizes for R, Rho, delta, and Delta (R_stepsize, Rho_stepsize, delta_stepsize); they're all 0.2 by default}

\item{verbose}{logical variable for printing progress bar. Default to FALSE.}
}
\value{
a dataframe with input arguments, posterior samples, Metropolis algorithm acceptance rates, etc
}
\description{
This is a function for running the Markov chain Monte Carlo algorithm for the BMVMR_POCov Model. The first six arguments are required.
}
\examples{
\dontrun{
data("cholesterol")
Outcome <- cbind(cholesterol$pldlc, cholesterol$phdlc, cholesterol$ptg)
SD <- cbind(cholesterol$sdldl, cholesterol$sdhdl, cholesterol$sdtg)
Trial <- cholesterol$Trial
Treat <- cholesterol$trt
Npt <- cholesterol$Npt
XCovariate <- cbind(cholesterol$bldlc, cholesterol$bhdlc, cholesterol$btg, cholesterol$age, cholesterol$durat, cholesterol$white, cholesterol$male, cholesterol$dm)
WCovariate <- cbind(1-cholesterol$onstat, cholesterol$trt * (1-cholesterol$onstat), cholesterol$onstat, cholesterol$trt * cholesterol$onstat)

fmodel <- 3
fit <- bayes.parobs(Outcome, SD, scale(XCovariate, scale=TRUE, center=TRUE), WCovariate, Treat, Trial, Npt, fmodel,
			mcmc=list(ndiscard=100000,nskip=1,nkeep=20000), verbose = TRUE)
}
}
\author{
Daeyoung Lim, \email{daeyoung.lim@uconn.edu}
}
