% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-cholesterol.R
\docType{data}
\name{cholesterol}
\alias{cholesterol}
\title{Cholesterol data}
\format{An object of class \code{data.frame} with 52 rows and 17 columns.}
\usage{
data(cholesterol)
}
\description{
Data collected by Merck & Co. for meta-analysis
}
\examples{
data(cholesterol)
}
\keyword{datasets}
