% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.bsynthesis}
\alias{coef.bsynthesis}
\title{get the posterior mean of fixed-effect coefficients}
\usage{
\method{coef}{bsynthesis}(object, ...)
}
\arguments{
\item{object}{a class of \code{bsynthesize}}

\item{...}{other arguments}
}
\value{
Coefficients extracted from the model object \code{object}
}
\description{
get the posterior mean of fixed-effect coefficients
}
