% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpd.bayesnmr.R
\name{confint.bayesnmr}
\alias{confint.bayesnmr}
\title{get the highest posterior density (HPD) interval}
\usage{
\method{confint}{bayesnmr}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the probability which the HPD interval will cover}

\item{...}{additional arguments for confint}
}
\value{
dataframe containing HPD intervals for the parameters
}
\description{
get the highest posterior density (HPD) interval
}
