% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-data.R
\docType{data}
\name{df}
\alias{df}
\title{Network Meta Regression data for Bayesian Network Meta-Regression Hierarchical Models Using Heavy-Tailed Multivariate Random Effects with Covariate-Dependent Variances}
\format{An object of class \code{data.frame} with 90 rows and 10 columns.}
\usage{
data(df)
}
\description{
Data simulated from the model in the paper.
}
\examples{
data(df)
}
\keyword{datasets}
