% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.bayes.parobs.R
\name{fitted.bayes.parobs}
\alias{fitted.bayes.parobs}
\title{get fitted values}
\usage{
\method{fitted}{bayes.parobs}(object, conf.level = 0.95, HPD = TRUE,
  ...)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}
}
\value{
a list of fitted values
}
\description{
get fitted values
}
