% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.bayesnmr.R
\name{fitted.bayesnmr}
\alias{fitted.bayesnmr}
\title{get fitted values}
\usage{
\method{fitted}{bayesnmr}(object, conf.level = 0.95, HPD = TRUE, ...)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}
}
\value{
a list of fitted values
}
\description{
get fitted values
}
