% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.bayes.parobs.R
\name{gof.bayes.parobs}
\alias{gof.bayes.parobs}
\title{get goodness of fit}
\usage{
\method{gof}{bayes.parobs}(object, type = "lpml", ncores = NULL, verbose = FALSE)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}

\item{type}{the type of goodness of fit to compute; DIC or LPML}

\item{ncores}{the number of CPU cores to use for parallel processing; it must not exceed the number of existing cores}

\item{verbose}{FALSE by default; If TRUE, then progress bar will appear}
}
\description{
get goodness of fit
}
