% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\title{get goodness of fit: DIC or LPML}
\usage{
gof(object, type = "lpml", verbose = FALSE)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}

\item{type}{the type of goodness of fit to compute; DIC or LPML}

\item{verbose}{FALSE by default; If TRUE, then progress bar will appear}
}
\value{
dataframe containing the goodness of fit measure
}
\description{
get goodness of fit: DIC or LPML
}
