% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpd.bayesparobs.R
\name{hpd.bayesparobs}
\alias{hpd.bayesparobs}
\title{get the highest posterior density (HPD) interval or equal-tailed credible interval}
\usage{
\method{hpd}{bayesparobs}(object, parm, level = 0.95, HPD = TRUE)
}
\arguments{
\item{object}{the output model from fitting a (network) meta analysis/regression model}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the probability which the HPD interval will cover}

\item{HPD}{a logical value indicating whether HPD or equal-tailed credible interval should be computed; by default, TRUE}
}
\value{
dataframe containing HPD intervals for the parameters
}
\description{
get the highest posterior density (HPD) interval or equal-tailed credible interval
}
