% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_cor.R
\name{meta_matrix_cor}
\alias{meta_matrix_cor}
\title{\code{meta_matrix_cor} fits the meta-regression model where the correlation matrices are covariate-modeled}
\usage{
meta_matrix_cor(
  Outcome,
  SD,
  XCovariate,
  WCovariate,
  ZCovariate,
  Treat,
  Trial,
  Npt,
  transform_type = 1,
  prior = list(),
  mcmc = list(),
  control = list(),
  init = list(),
  Treat_order = NULL,
  Trial_order = NULL,
  group = NULL,
  group_order = NULL,
  scale_x = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Outcome}{the aggregate mean of the responses for each arm of every study.}

\item{SD}{the standard deviation of the responses for each arm of every study.}

\item{XCovariate}{the aggregate covariates for the fixed effects.}

\item{WCovariate}{the aggregate covariates for the random effects.}

\item{ZCovariate}{the aggregate covariates for modeling the correlation matrices.}

\item{Treat}{the treatment identifiers. This is equivalent to the arm number of each study. The number of unique treatments must be equal across trials. The elements within will be coerced to consecutive integers.}

\item{Trial}{the trial identifiers. This is equivalent to the arm labels in each study. The elements within will be coerced to consecutive integers}

\item{Npt}{the number of observations/participants for a unique \verb{(t,k)}, or each arm of every trial.}

\item{transform_type}{the model number. The possible values for \code{transform_type} are 1 to 3, each indicating a different prior specification for \eqn{\rho_{tk}}. It will default to M1, \code{transform_type=1} if not specified at function call. See the following model descriptions.
\itemize{
\item \code{transform_type=1} - \eqn{\rho_{tk}} is modeled as a matrix exponential of a real symmetric matrix whose upper-diagonal elements are unconstrained and modeled as a determistic function of the covariates.
\item \code{transform_type=2} - \eqn{\rho_{tk}} is modeled as a correlation matrix whose partial correlation matrix is Fisher's z-transformed so the upper-diagonal elements are unconstrained. The free variables are then modeled as a determistic function of the covariates.
\item \code{transform_type=3} - \eqn{\rho_{tk}} is modeled as a correlation matrix whose Cholesky factor is modeled as hyperspherical coordinates. The hyperspherical coordinates are a deterministic function of the covariates.
}}

\item{prior}{(Optional) a list of hyperparameters. Despite \code{theta} in every model, each \code{fmodel}, along with the \code{group} argument, requires a different set of hyperparameters.}

\item{mcmc}{(Optional) a list for MCMC specification. \code{ndiscard} is the number of burn-in iterations. \code{nskip} configures the thinning of the MCMC. For instance, if \code{nskip=5}, \code{bayes.parobs} will save the posterior sample every 5 iterations. \code{nkeep} is the size of the posterior sample. The total number of iterations will be \code{ndiscard + nskip * nkeep}.}

\item{control}{(Optional) a list of tuning parameters for \href{https://en.wikipedia.org/wiki/Metropolis-Hastings_algorithm}{the Metropolis-Hastings algorithm}. \code{R}, and \code{delta} are sampled through either localized Metropolis algorithm or delayed rejection robust adaptive Metropolis algorithm. \verb{*_stepsize} with the asterisk replaced with one of the names above specifies the stepsize for determining the sample evaluation points in the localized Metropolis algorithm.}

\item{init}{(Optional) a list of initial values for the parameters to be sampled: \code{theta}, \code{gamR}, and \code{Omega}.}

\item{Treat_order}{(Optional) a vector of unique treatments to be used for renumbering the \code{Treat} vector. The first element will be assigned treatment zero, potentially indicating placebo. If not provided, the numbering will default to an alphabetical/numerical order.}

\item{Trial_order}{(Optional) a vector of unique trials. The first element will be assigned zero. If not provided, the numbering will default to an alphabetical/numerical order.}

\item{group}{(Optional) a vector containing binary variables for \eqn{u_{tk}}. If not provided, \code{bayes.parobs} will assume that there is no grouping and set \eqn{u_{tk}=0} for all \verb{(t,k)}.}

\item{group_order}{(Optional) a vector of unique group labels. The first element will be assigned zero. If not provided, the numbering will default to an alphabetical/numerical order. \code{group_order} will take effect only if \code{group} is provided by the user.}

\item{scale_x}{(Optional) a logical variable indicating whether \code{XCovariate} should be scaled/standardized. The effect of setting this to \code{TRUE} is not limited to merely standardizing \code{XCovariate}. The following generic functions will scale the posterior sample of \code{theta} back to its original unit: \code{plot}, \code{fitted}, \code{summary}, and \code{print}.}

\item{verbose}{(Optional) a logical variable indicating whether to print the progress bar during the MCMC sampling.}
}
\description{
\code{meta_matrix_cor} extends the meta-regression model so that the correlation matrix can be modeled using covariates. There are three options for modeling the correlation matrix: (1) matrix exponential, (2) partial correlation, and (3) Cholesky factor further transformed into hyperspherical coordinates. Note that the latter two are not \emph{permutation invariant}. This means that \eqn{(y_1, y_2, y_3)} and \eqn{(y_2, y_1, y_3)} (or any permutation of the random variables) will have a nonsignificant effect on the estimation and the statistical inference.
}
\examples{
set.seed(2797542)
data("cholesterol")
f_1 <- 'pldlc + phdlc + ptg | sdldl + sdhdl + sdtg ~ 0 + bldlc + bhdlc + btg +
  age + durat + white + male + dm + ns(n) | treat | treat + trial + onstat'
out_1 <- bmeta_analyze(as.formula(f_1), data = cholesterol,
  prior = list(model="NoRecovery"),
  mcmc = list(ndiscard = 3, nskip = 1, nkeep = 1),
  control=list(scale_x = TRUE, verbose=FALSE))

set.seed(2797542)
data("TNM")
TNM$group <- factor(match(TNM$treat, c("PBO", "R"), nomatch = 0))
f_2 <- 'ptg | sdtg ~
  0 + bldlc + bhdlc + btg + age + white + male + bmi +
  potencymed + potencyhigh + durat + ns(n) |
  scale(bldlc) + scale(btg) + group | treat  + trial'
out_2 <- bmeta_analyze(as.formula(f_2), data = TNM,
  mcmc = list(ndiscard = 1, nskip = 1, nkeep = 1),
  control=list(scale_x = TRUE, verbose=FALSE))
}
\references{
Yao, H., Kim, S., Chen, M. H., Ibrahim, J. G., Shah, A. K., & Lin, J. (2015). Bayesian inference for multivariate meta-regression with a partially observed within-study sample covariance matrix. \emph{Journal of the American Statistical Association}, \strong{110(510)}, 528-544.

Li, H., Chen, M. H., Ibrahim, J. G., Kim, S., Shah, A. K., Lin, J., & Tershakovec, A. M. (2019). Bayesian inference for network meta-regression using multivariate random effects with applications to cholesterol lowering drugs. \emph{Biostatistics}, \strong{20(3)}, 499-516.

Li, H., Lim, D., Chen, M. H., Ibrahim, J. G., Kim, S., Shah, A. K., & Lin, J. (2021). Bayesian network meta-regression hierarchical models using heavy-tailed multivariate random effects with covariate-dependent variances. \emph{Statistics in Medicine}.
}
\author{
Daeyoung Lim, \email{daeyoung.lim@uconn.edu}
}
