\name{metapack-package}
\alias{metapack-package}
\alias{metapack}
\docType{package}
\title{
\packageTitle{metapack}
}
\description{
\packageDescription{metapack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{metapack}
\packageIndices{metapack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{metapack}

Maintainer: \packageMaintainer{metapack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
