% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.comp.R
\name{model.comp}
\alias{model.comp}
\title{compute the model comparison measures: DIC or LPML}
\usage{
model.comp(object, type = "lpml", verbose = FALSE, ncores = NULL)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}

\item{type}{the type of model comparison measure to compute; DIC or LPML}

\item{verbose}{FALSE by default; If TRUE, then progress bar will appear}

\item{ncores}{the number of CPU cores to use for parallel processing; it must not exceed the number of existing cores}
}
\value{
dataframe containing the compute the model comparison measures
}
\description{
compute the model comparison measures: DIC or LPML
}
