% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comp.bayesnmr.R
\name{model_comp.bayesnmr}
\alias{model_comp.bayesnmr}
\title{get compute the model comparison measures}
\usage{
\method{model_comp}{bayesnmr}(object, type = "lpml", verbose = FALSE, ncores = NULL)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}

\item{type}{the type of model comparison measures; DIC or LPML}

\item{verbose}{FALSE by default; If TRUE, then progress bar will appear}

\item{ncores}{the number of CPU cores to use for parallel processing. It must not exceed the number of existing cores. If unspecified, it will default to 2 cores or the number of existing cores, whichever is smaller.}
}
\value{
dataframe containing the compute the model comparison measures
}
\description{
get compute the model comparison measures
}
