% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ns}
\alias{ns}
\title{helper function encoding trial sample sizes in formulas}
\usage{
ns(x)
}
\arguments{
\item{x}{the name of the variable containing trial sample sizes}
}
\description{
helper function encoding trial sample sizes in formulas
}
