% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bayesnmr.R
\name{plot.bayesnmr}
\alias{plot.bayesnmr}
\title{get goodness of fit}
\usage{
\method{plot}{bayesnmr}(x, ...)
}
\arguments{
\item{x}{the output model from fitting a meta analysis/regression model}

\item{...}{additional parameters for plot}
}
\value{
No return value
}
\description{
get goodness of fit
}
