% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bayes.parobs.R
\name{print.bayes.parobs}
\alias{print.bayes.parobs}
\title{Print results}
\usage{
\method{print}{bayes.parobs}(object, conf.level = 0.95, HPD = TRUE,
  ...)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}
}
\value{
does not return anything; print a summary of the output
}
\description{
Print results
}
