% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bayesnmr.R
\name{print.bayesnmr}
\alias{print.bayesnmr}
\title{Print results}
\usage{
\method{print}{bayesnmr}(object, conf.level = 0.95, HPD = TRUE, ...)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}
}
\value{
does not return anything; print a summary of the output
}
\description{
Print results
}
