% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bayesparobs.R
\name{print.bayesparobs}
\alias{print.bayesparobs}
\title{Print results}
\usage{
\method{print}{bayesparobs}(x, level = 0.95, HPD = TRUE, ...)
}
\arguments{
\item{x}{the output model from fitting a meta analysis/regression model}

\item{level}{credible level for interval estimation; set to 0.95 by default}

\item{HPD}{a logical argument indicating whether HPD intervals should be computed; if FALSE, equal-tail credible intervals are computed}

\item{...}{additional arguments for print}
}
\value{
No return value; print a summary of the output
}
\description{
Print results
}
