% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sucra.bayesnmr.R
\name{sucra.bayesnmr}
\alias{sucra.bayesnmr}
\title{get surface under the cumulative ranking curve (SUCRA)}
\usage{
\method{sucra}{bayesnmr}(object)
}
\arguments{
\item{object}{the output model from fitting a network meta analysis/regression model}
}
\value{
a list containing SUCRA and the discrete rank probability matrix of size T by T
}
\description{
get surface under the cumulative ranking curve (SUCRA)
}
