% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bayes.parobs.R
\name{summary.bayes.parobs}
\alias{summary.bayes.parobs}
\title{\code{summary} method for class "\code{bayes.parobs}"}
\usage{
\method{summary}{bayes.parobs}(object, ...)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}

\item{...}{additional arguments for summary}
}
\value{
print summary for the model fit
}
\description{
\code{summary} method for class "\code{bayes.parobs}"
}
