% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bayesparobs.R
\name{summary.bayesparobs}
\alias{summary.bayesparobs}
\title{\code{summary} method for class "\code{bayesparobs}"}
\usage{
\method{summary}{bayesparobs}(object, level = 0.95, HPD = TRUE, ...)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}

\item{level}{credible level for interval estimation; set to 0.95 by default}

\item{HPD}{a logical argument indicating whether HPD intervals should be computed; if FALSE, equal-tail credible intervals are computed}

\item{...}{additional arguments for summary}
}
\value{
print summary for the model fit
}
\description{
\code{summary} method for class "\code{bayesparobs}"
}
