% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize.R
\name{synthesize}
\alias{synthesize}
\title{Wrapper function for worker functions: bayes.parobs, bayes.nmr}
\usage{
synthesize(formula, data, prior = list(), control = list(), ...)
}
\arguments{
\item{formula}{an object of class of \link[Formula]{Formula}: a symbolic description of the meta-analytic model to be fitted. The list of models includes multivariate meta-regression and univariate network meta-regression. More will be added moving forward.}

\item{data}{a data frame, list, or environment (or object coercible by \link[base]{as.data.frame} to a data frame) containing the variables in the model. If not found in `data`, the variables are taken from `environment(formula)`, typically the environment from which `synthesize` is called.}

\item{prior}{an optional object that contains the hyperparameter values for the model}

\item{control}{an optional object that contains the control tuning parameters for the Metropolis-Hastings algorithm}
}
\description{
This is the one function to rule them all. All other worker functions will be subsumed by this function, so that users can forget about the implementation details.
}
\author{
Daeyoung Lim, \email{daeyoung.lim@uconn.edu}
}
