% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Fit test function}
\usage{
test(X, ndiscard, nskip, nkeep)
}
\arguments{
\item{X}{a matrix of data whose columns are X_i}

\item{ndiscard}{number of burn-in iterations}

\item{nskip}{number of thinning iterations}

\item{nkeep}{number of posterior samples}
}
\value{
a dataframe with input arguments, posterior samples, Metropolis algorithm acceptance rates, etc
}
\description{
This is a function for testing the Metropolis-Hastings algorithm
}
\examples{
\dontrun{
data("cholesterol")
Outcome <- cbind(cholesterol$pldlc, cholesterol$phdlc, cholesterol$ptg)
SD <- cbind(cholesterol$sdldl, cholesterol$sdhdl, cholesterol$sdtg)
Trial <- cholesterol$Trial
Treat <- cholesterol$trt
Npt <- cholesterol$Npt
XCovariate <- cbind(cholesterol$bldlc, cholesterol$bhdlc, cholesterol$btg, cholesterol$age, cholesterol$durat, cholesterol$white, cholesterol$male, cholesterol$dm)
WCovariate <- cbind(1-cholesterol$onstat, cholesterol$trt * (1-cholesterol$onstat), cholesterol$onstat, cholesterol$trt * cholesterol$onstat)

fmodel <- 3
fit <- bayes.parobs(Outcome, SD, scale(XCovariate, scale=TRUE, center=TRUE), WCovariate, Treat, Trial, Npt, fmodel,
			mcmc=list(ndiscard=100000,nskip=1,nkeep=20000), control(delta_rep=100, rho_rep=200), verbose = TRUE)
}
}
\author{
Daeyoung Lim, \email{daeyoung.lim@uconn.edu}
}
