% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.parobs.R
\name{bayes.parobs}
\alias{bayes.parobs}
\title{Fit Bayesian Inference for Multivariate Meta-Regression With a Partially Observed Within-Study Sample Covariance Matrix}
\usage{
bayes.parobs(Outcome, SD, XCovariate, WCovariate, Treat, Trial, Npt,
  fmodel = 1, prior = list(), mcmc = list(), verbose = FALSE)
}
\arguments{
\item{Outcome}{aggregate mean of the responses for each arm of each study}

\item{SD}{standard deviation of the responses for each arm of each study}

\item{XCovariate}{aggregate covariates for the fixed effects}

\item{WCovariate}{aggregate covariates for the random effects}

\item{Treat}{treatment applied; equivalent to the arm number of each study; the number of unique treatments must be equal across trials}

\item{Trial}{trial identifier}

\item{Npt}{number of observations per trial}

\item{fmodel}{the model number; defaults to M1}

\item{prior}{list of hyperparameters; when not given, algorithm will run in default setting}

\item{mcmc}{list of MCMC-related parameters: number of burn-ins (ndiscard), number of thinning(nskip), and posterior sample size (nkeep)}

\item{verbose}{logical variable for printing progress bar. Default to FALSE.}
}
\value{
a dataframe with input arguments, posterior samples, Metropolis algorithm acceptance rates, etc
}
\description{
This is a function for running the Markov chain Monte Carlo algorithm for the BMVMR_POCov Model. The first six arguments are required.
}
\examples{
\dontrun{
data("cholesterol")
Outcome <- cbind(cholesterol$ldlcm, cholesterol$hdlcm, cholesterol$tgm)
SD <- cbind(cholesterol$ldlcsd, cholesterol$hdlcsd, cholesterol$tgsd)
Trial <- cholesterol$Trial
Treat <- cholesterol$Trt
Npt <- cholesterol$npt
XCovariate <- cbind(cholesterol$bl_ldlc, cholesterol$bl_hdlc, cholesterol$bl_tg, cholesterol$age, cholesterol$Dur, cholesterol$white, cholesterol$male, cholesterol$DM)
WCovariate <- cbind(1, cholesterol$Trt)

fit <- bayes.parobs(Outcome, SD, XCovariate, WCovariate, Treat, Trial, Npt, 1,
			mcmc=list(ndiscard=2500,nskip=1,nkeep=10000), verbose = TRUE)
}
}
\author{
Daeyoung Lim, \email{daeyoung.lim@uconn.edu}
}
