% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.bayesnmr.R
\name{fitted.bayesnmr}
\alias{fitted.bayesnmr}
\title{get fitted values}
\usage{
\method{fitted}{bayesnmr}(object, level = 0.95, HPD = TRUE, ...)
}
\arguments{
\item{object}{the output model from fitting a meta analysis/regression model}

\item{level}{credible level for interval estimation; set to 0.95 by default}

\item{HPD}{a logical argument indicating whether HPD intervals should be computed; if FALSE, equal-tail credible intervals are computed}

\item{...}{additional arguments for fitted}
}
\value{
a list of fitted values
}
\description{
get fitted values
}
