% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sucra.R
\name{plot.sucra}
\alias{plot.sucra}
\title{plot the surface under the cumulative ranking curve (SUCRA)}
\usage{
\method{plot}{sucra}(object, ...)
}
\arguments{
\item{object}{the output model from fitting a network meta analysis/regression model}
}
\description{
plot the surface under the cumulative ranking curve (SUCRA)
}
