% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bayesnmr.R
\name{summary.bayesnmr}
\alias{summary.bayesnmr}
\title{Summarize results}
\usage{
\method{summary}{bayesnmr}(object, ...)
}
\arguments{
\item{object}{the output model from fitting a network meta analysis/regression model}

\item{...}{additional arguments for print}
}
\value{
does not return anything; print a summary of the output
}
\description{
Summarize results
}
